refactor: VoteRequest, Response 통합된 클래스로 관리
