+++ src/main/java/balancetalk/vote/presentation/VoteGameController.java
@@ -1,7 +1,9 @@
 package balancetalk.vote.presentation;
 
-import balancetalk.vote.dto.VoteRequest;
-import balancetalk.vote.dto.VoteResponse;
+import static balancetalk.vote.dto.VoteDto.*;
+
+import balancetalk.vote.dto.VoteDto;
+import balancetalk.vote.dto.VoteDto.VoteResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.web.bind.annotation.GetMapping;
